assets/sources/cosmosdb_reader/requirements.txt
dagster
pandas>=1.5.0
azure-cosmos>=4.0.0
